+++ data/tables/advancment/hero.js
@@ -12,7 +12,7 @@ module.exports = {
   entries: [{
     lowerBounds: 2,
     upperBounds: 5,
-    text: 'New Skill'
+    text: () => 'New Skill'
   },
   {
     lowerBounds: 6,
@@ -25,12 +25,12 @@ module.exports = {
         {
           lowerBounds: 1,
           upperBounds: 3,
-          text: '+1 Strength (+1 Attack if already at max Strength)'
+          text: () => '+1 Strength (+1 Attack if already at max Strength)'
         },
         {
           lowerBounds: 4,
           upperBounds: 6,
-          text: '+1 Attack (+1 Strength if already at max Attack)'
+          text: () => '+1 Attack (+1 Strength if already at max Attack)'
         },
       ]
     }
@@ -38,7 +38,7 @@ module.exports = {
   {
     lowerBounds: 7,
     upperBounds: 7,
-    text: 'Choose either +1WS or +1BS'
+    text: () => 'Choose either +1WS or +1BS'
   },
   {
     lowerBounds: 8,
@@ -51,12 +51,12 @@ module.exports = {
         {
           lowerBounds: 1,
           upperBounds: 3,
-          text: '+1 Inititave (+1 Leadership if already at max Inititave)'
+          text: () => '+1 Inititave (+1 Leadership if already at max Inititave)'
         },
         {
           lowerBounds: 4,
           upperBounds: 6,
-          text: '+1 Leadership (+1 Inititave if already at max Leadership)'
+          text: () => '+1 Leadership (+1 Inititave if already at max Leadership)'
         },
       ]
     }
@@ -72,12 +72,12 @@ module.exports = {
         {
           lowerBounds: 1,
           upperBounds: 3,
-          text: '+1 Wound (+1 Toughness if already at max Wounds)'
+          text: () => '+1 Wound (+1 Toughness if already at max Wounds)'
         },
         {
           lowerBounds: 4,
           upperBounds: 6,
-          text: '+1 Toughness (+1 Wound if already at max Toughness)'
+          text: () => '+1 Toughness (+1 Wound if already at max Toughness)'
         },
       ]
     }
@@ -85,7 +85,7 @@ module.exports = {
   {
     lowerBounds: 10,
     upperBounds: 12,
-    text: 'New Skill'
+    text: () => 'New Skill'
   },
   ]
-}
+};