+++ data/tables/tables.js
@@ -1,10 +1,12 @@
 const advhero = require('./advancment/hero');
 const advhench = require('./advancment/henchmen');
+const injury = require('./injury/injury')
 
 
 const Tables = {
   advhero,
-  advhench
+  advhench,
+  injury
 }
 
 module.exports = Tables