+++ CHANGELOG.md
@@ -9,3 +9,11 @@
 - Hero Advancement table
 - Henchmen Advancement table
 - Faction Management Feature
+
+## [1.0.1] - 14-06-2024
+
+### Added
+- Injury table 
+- Unassigned populations
+- Population total
+- Shortcuts for /strangeroll and /strangetable commands. These commands can now be accessed with /st and /sr eg. /sr 2d6